backend/requirements.txt

celery==5.3.4
channels>=4.0,<5
channels-redis>=4.1
Django==4.2.7
django-cors-headers==4.3.1
django-crispy-forms==2.4
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
redis==5.0.1
reportlab==4.4.9
social-auth-app-django==5.5.0


